Added some explanatory comments
