+++ packages/core/package.json
@@ -14,7 +14,9 @@
   "dependencies": {
     "graphql": "^15.3.0",
     "handlebars": "^4.7.6",
+    "human-id": "^2.0.1",
     "immer": "7.0.14",
-    "human-id": "^2.0.1"
+    "lodash.camelcase": "^4.3.0",
+    "lodash.upperfirst": "^4.3.1"
   }
 }
