+++ packages/core/src/createSextantPlugin.ts
@@ -1,9 +1,16 @@
 import { flattenDatabase } from './databaseUtilities';
-import { Database, FlattenedDatabase } from './types';
+import {
+  Database,
+  FlattenedDatabase,
+  SextantPluginImplementation,
+} from './types';
 import * as fs from 'fs';
 import * as path from 'path';
 import { getTargetDir } from './fileSaveUtilities';
 
+/**
+ * The context passed down to your plugin implementation
+ */
 export interface SextantContext {
   /**
    * The entire database, for your perusal
@@ -16,16 +23,10 @@ export interface SextantContext {
   writeFileSync: (relativePath: string, contents: string) => void;
 }
 
-export type SextantPlugin = <TConfig extends {}>(
-  database: Database,
-  config: TConfig,
-) => void;
-
-export type SextantPluginImplementation<TConfig extends {}> = (
-  sextantContext: SextantContext,
-  config: TConfig,
-) => void;
-
+/**
+ * Creates a plugin which can be used to print files
+ * based on the sextant database
+ */
 export const createSextantPlugin = <TConfig extends {}>(
   implementation: SextantPluginImplementation<TConfig>,
 ) =>
