+++ packages/plugin-typescript/package.json
@@ -17,8 +17,8 @@
     "fs-extra": "^9.0.1",
     "graphql": "^15.4.0",
     "handlebars": "^4.7.6",
-    "handlebars-helpers": "^0.10.0",
-    "lodash": "^4.17.20"
+    "lodash.camelcase": "^4.3.0",
+    "lodash.upperfirst": "^4.3.1"
   },
   "devDependencies": {
     "@types/fs-extra": "^9.0.3"
